requirements.txt
pexpect
requests
stevedore
PyYaml
Jinja2
